+++ src/components/BlogRoll.js
@@ -4,26 +4,17 @@ import { Link, graphql, StaticQuery } from 'gatsby'
 import '../styles/index.css'
 import Layout from './Layout.js'
 import { getImage, GatsbyImage } from 'gatsby-plugin-image'
-import '../styles/index.css'
+import WhichPost from './WhichPost'
+
 function BlogRoll({data}){
   const posts = data.allMarkdownRemark.edges
-  let pages = [0];
-  let numberOfPage = 1
-  const howManyPages = ()=>{
-    for (let i = 0; i < posts.length; i++){
-      if((i+1)%4 == 0){
-        pages.push(numberOfPage)
-        numberOfPage++
-      }
-    }
-    return pages
-  }
+  console.log(posts)
   const shouldShow = (index)=>{
+    if(index < 4 + 4*localStorage.getItem('whichPage') && index > 4*localStorage.getItem('whichPage')-1){
+      localStorage.clear()
       return true
-  }
-  function changePage(index){
-    console.log('button clicked' + index)
-  }
+    }
+}
   return(
     <Layout>
       <div className='blog-list__content-wrapper'>
@@ -31,15 +22,16 @@ function BlogRoll({data}){
           posts.map(({node:post},index)=>(
             post.frontmatter.mainImage&&shouldShow(index)?(
               <div key={post.id} className='blog-list__wrapper'>
-                {/* <div>{index}</div> */}
                 <GatsbyImage className='blog-list blog-list--picture' image={
                   getImage(post.frontmatter.mainImage)
                 } alt='mainimage'/>
                 <div className='blog-list blog-list--description'>
                   <p className='blog-list blog-list--date'>
-                    {post.frontmatter.date}</p>
+                    {post.frontmatter.date}
+                  </p>
                   <p className='blog-list blog-list--title'>
-                    {post.frontmatter.title}</p>
+                    {post.frontmatter.title}
+                  </p>
                   <p className='blog-list blog-list--link'>
                     <Link to={post.frontmatter.path}>
                       kliknij tu!
@@ -50,17 +42,7 @@ function BlogRoll({data}){
             ):''
           ))
         }
-        <div className='blog-list__other-posts'>
-          <span>poprzednia</span>
-            {
-              howManyPages().map((element,index)=>(
-                <button key={index} onClick={()=>changePage(element)}>
-                  {element}
-                </button>
-              ))
-            }
-          <span>następna</span>
-        </div>
+        <WhichPost posts={posts}/>
       </div>
     </Layout>)
 }
@@ -81,7 +63,6 @@ const query = () => (
         ) {
           edges {
             node {
-              excerpt(pruneLength: 400)
               id
               frontmatter {
                 path
