+++ src/pages/index.js
@@ -2,12 +2,7 @@ import React from 'react'
 // import { Link } from 'gatsby'
 import BlogRoll from '../components/BlogRoll'
 const index = () => {
-  return(
-    <>
-      <BlogRoll/>
-      
-    </>
-  )
+  return <BlogRoll/>
 }
 
 export default index
